refactor: Separando os comportamentos e dados das chaves do dominio do servidor.
